+++ CHANGELOG.md
@@ -5,10 +5,16 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project does adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.4.0] - Unreleased
+Added:
+- Progress bar for extract
+- "Extract all" from within QuickLook preview
+
+
 ## [1.3.0] – 2026-04-15
 Added:
 - Settings popover for default options
-- Export all action
+- "Extract all" action
 - Resolve symlinks action
 
 Changed:
@@ -55,6 +61,7 @@ Fixed:
 Initial release
 
 
+[1.4.0]: https://github.com/relikd/QLArchives/compare/v1.3.0...HEAD
 [1.3.0]: https://github.com/relikd/QLArchives/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/relikd/QLArchives/compare/v1.1.1...v1.2.0
 [1.1.1]: https://github.com/relikd/QLArchives/compare/v1.1.0...v1.1.1